# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_pranavbh-9117_IMB_internal_attempt_dto.AnswerSubmission:
    properties:
      question_id:
        type: string
      selected_option_id:
        description: Nullable if the student skips the question
        type: string
    required:
    - question_id
    type: object
  github_com_pranavbh-9117_IMB_internal_attempt_dto.FacultyResultResponse:
    properties:
      attempt_id:
        type: string
      score:
        type: integer
      started_at:
        type: string
      student_email:
        type: string
      student_id:
        type: string
      student_name:
        type: string
      submitted_at:
        type: string
      total_marks:
        type: integer
    type: object
  github_com_pranavbh-9117_IMB_internal_attempt_dto.StudentResultResponse:
    properties:
      attempt_id:
        type: string
      quiz_id:
        type: string
      quiz_title:
        type: string
      score:
        type: integer
      started_at:
        type: string
      submitted_at:
        type: string
      total_marks:
        type: integer
    type: object
  github_com_pranavbh-9117_IMB_internal_attempt_dto.SubmitAttemptRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_attempt_dto.AnswerSubmission'
        type: array
    required:
    - answers
    type: object
  github_com_pranavbh-9117_IMB_internal_auth_dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  github_com_pranavbh-9117_IMB_internal_auth_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_pranavbh-9117_IMB_internal_auth_dto.LoginResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_auth_dto.UserResponse'
    type: object
  github_com_pranavbh-9117_IMB_internal_auth_dto.RefreshResponse:
    properties:
      access_token:
        type: string
    type: object
  github_com_pranavbh-9117_IMB_internal_auth_dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse:
    properties:
      data: {}
      message:
        example: error description
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any:
    properties:
      data: {}
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  ? github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_attempt_dto_FacultyResultResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_attempt_dto.FacultyResultResponse'
        type: array
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  ? github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_attempt_dto_StudentResultResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_attempt_dto.StudentResultResponse'
        type: array
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_auth_dto_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_auth_dto.LoginResponse'
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_auth_dto_RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_auth_dto.RefreshResponse'
      message:
        example: operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the Institute Management Platform.
  title: Institute Management Backend API
  version: "1.0"
paths:
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password and revokes all active
        sessions.
      parameters:
      - description: Change Password Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_auth_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password Changed Successfully
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Change Password
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT access token in the body
        and a refresh token in an HttpOnly cookie.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_auth_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Successful
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_auth_dto_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Revokes the refresh token and clears the HttpOnly cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Logout Successful
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      summary: User Logout
      tags:
      - Authentication
  /auth/refresh:
    post:
      description: Reads the refresh token from the HttpOnly cookie, rotates it, and
        returns a new access token.
      produces:
      - application/json
      responses:
        "200":
          description: Token Refreshed
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-github_com_pranavbh-9117_IMB_internal_auth_dto_RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      summary: Refresh Token
      tags:
      - Authentication
...
  /quizzes/{id}/attempt:
    post:
      consumes:
      - application/json
      description: Submits a quiz attempt. Students only. Quiz must be published and
        student must belong to the same institution. One attempt per student.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      - description: Attempt Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_pranavbh-9117_IMB_internal_attempt_dto.SubmitAttemptRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attempt Submitted
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit Quiz Attempt
      tags:
      - Quiz Attempts
...
  /quizzes/{id}/results:
    get:
      description: Retrieves performance data for a specific quiz. Faculty only. Can
        only view results for quizzes they created.
      parameters:
      - description: Quiz ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quiz Results Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_attempt_dto_FacultyResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Quiz Results
      tags:
      - Quiz Attempts
  /results:
    get:
      description: Retrieves all quiz results for the authenticated student.
      produces:
      - application/json
      responses:
        "200":
          description: Results Retrieved
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerResponse-array_github_com_pranavbh-9117_IMB_internal_attempt_dto_StudentResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pranavbh-9117_IMB_pkg_response.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Student Results
      tags:
      - Quiz Attempts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"